+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.proton` artifact to version `0.68`.
+
 ### [v0.219](https://github.com/react4j/react4j/tree/v0.219) (2026-04-29) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.218...v0.219)
 
 Changes in this release: